Revert "Refactor run command to use setup_repo from evaluation pipeline (#239)"
This reverts commit f2a65dd99ca0ad36e04c9bba0f22afb7f1b7c767.
