+++ src/bcbench/commands/run.py
@@ -19,7 +19,8 @@ from bcbench.cli_options import (
 from bcbench.config import get_config
 from bcbench.dataset import DatasetEntry, load_dataset_entries
 from bcbench.logger import get_logger
-from bcbench.operations import setup_repo
+from bcbench.operations import checkout_commit, clean_repo
+from bcbench.operations.instruction_operations import copy_problem_statement_folder
 
 logger = get_logger(__name__)
 _config = get_config()
@@ -46,7 +47,9 @@ def run_mini(
     """
     entry: DatasetEntry = load_dataset_entries(dataset_path, entry_id=entry_id)[0]
 
-    setup_repo(entry, repo_path, category)
+    clean_repo(repo_path)
+    checkout_commit(repo_path, entry.base_commit)
+    copy_problem_statement_folder(entry, repo_path)
 
     run_mini_agent(
         entry=entry,
@@ -77,7 +80,9 @@ def run_copilot(
     """
     entry: DatasetEntry = load_dataset_entries(dataset_path, entry_id=entry_id)[0]
 
-    setup_repo(entry, repo_path, category)
+    clean_repo(repo_path)
+    checkout_commit(repo_path, entry.base_commit)
+    copy_problem_statement_folder(entry, repo_path)
 
     run_copilot_agent(entry=entry, repo_path=repo_path, model=model, category=category, output_dir=output_dir, al_mcp=al_mcp)
