+++ src/bcbench/evaluate/testgeneration.py
@@ -1,6 +1,10 @@
 from collections.abc import Callable
+from pathlib import Path
+
+import yaml
 
 from bcbench.collection.patch_utils import extract_file_paths_from_patch
+from bcbench.config import get_config
 from bcbench.dataset import TestEntry
 from bcbench.evaluate.base import EvaluationPipeline
 from bcbench.exceptions import BuildError, NoTestsExtractedError, TestExecutionError
@@ -9,9 +13,11 @@ from bcbench.operations import (
     apply_patch,
     build_and_publish_projects,
     categorize_projects,
+    checkout_commit,
     clean_project_paths,
+    clean_repo,
+    copy_problem_statement_folder,
     extract_tests_from_patch,
-    setup_repo,
     stage_and_get_diff,
 )
 from bcbench.operations.bc_operations import run_test_suite
@@ -19,8 +25,29 @@ from bcbench.results.testgeneration import TestGenerationResult
 from bcbench.types import EvaluationContext
 
 logger = get_logger(__name__)
+_config = get_config()
 
 __all__ = ["TestGenerationPipeline"]
+
+
+def _get_test_generation_input_mode() -> str:
+    """Read test-generation input mode from copilot config.
+
+    Returns:
+        str: The validated input mode, either "gold-patch" or "problem-statement"
+
+    Raises:
+        ValueError: If the input mode is not one of the valid values
+    """
+    config_file: Path = _config.paths.agent_dir / "config.yaml"
+    copilot_config = yaml.safe_load(config_file.read_text())
+    input_mode = copilot_config.get("prompt", {}).get("test-generation-input", "problem-statement")
+
+    valid_modes = {"gold-patch", "problem-statement"}
+    if input_mode not in valid_modes:
+        raise ValueError(f"Invalid test-generation-input mode: '{input_mode}'. Must be one of {valid_modes}. Note: Use hyphens, not underscores (e.g., 'gold-patch' not 'gold_patch')")
+
+    return input_mode
 
 
 class TestGenerationPipeline(EvaluationPipeline):
@@ -37,7 +64,15 @@ class TestGenerationPipeline(EvaluationPipeline):
     """
 
     def setup(self, context: EvaluationContext) -> None:
-        setup_repo(context.entry, context.repo_path, context.category)
+        clean_repo(context.repo_path)
+        checkout_commit(context.repo_path, context.entry.base_commit)
+
+        input_mode: str = _get_test_generation_input_mode()
+        logger.info(f"Test generation input mode: {input_mode}")
+        if input_mode == "gold-patch":
+            apply_patch(context.repo_path, context.entry.patch, f"{context.entry.instance_id} gold patch")
+        else:
+            copy_problem_statement_folder(context.entry, context.repo_path)
 
         build_and_publish_projects(
             context.repo_path,
